THOLKAPIYAM/tolkappiyam-ai/requirements.txt
flask>=2.3.0
pandas>=2.0.0
networkx>=3.1
google-generativeai>=0.3.0
thol
python-dotenv>=1.0.0
reportlab>=4.0.0
